ec2-user@ec2-54-156-133-176.compute-1.amazonaws.com/RSVP/requirements.txt
django
djangorestframework
django-cors-headers
graphene-django
requests
pytest

